Prepare SDK for KLP MR1. Do not merge.
System images are in 19 rev 2 for MR1 (since they have minor.major)

platform-tools: update the API diff file that lint uses (not automated yet.)
build-tools: llvm-rs-cc issue affecting RS NDK mode in Gradle.

Note that support is set to 19.0.1 which means it will have to be
released using the new addon.xsd format that allows extras to have a
major.micro.minor format. That will be an interesting way to test this
feature.

Change-Id: Id755770f8a121edffa36e0fcb70488c452087440